split package tests
